refactor: cva variants 파일 분리
