+++ src/components/atoms/Input/Input.tsx
@@ -1,27 +1,7 @@
 import { InputHTMLAttributes, forwardRef } from "react";
-import { cva, type VariantProps } from "class-variance-authority";
+import { type VariantProps } from "class-variance-authority";
 import { cn } from "@/utils/cn";
-
-const inputVariants = cva(
-  "h-[24px] body-m-semibold focus:outline-none transition-colors caret-[#20A6FF]",
-  {
-    variants: {
-      variant: {
-        default:
-          "border-border-normal-lowemp focus:border-border-normal-highemp",
-        error:
-          "border-border-normal-highemp focus:border-border-normal-highemp",
-      },
-      fullWidth: {
-        true: "w-full",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-      fullWidth: false,
-    },
-  }
-);
+import { inputVariants } from "./inputVariants";
 
 interface InputProps
   extends InputHTMLAttributes<HTMLInputElement>,
